Further clarification on forgetting factors
In response to Abe Claycomb's technical algorithm review, I attempted to
further clarify how alpha, beta, and gamma work. This includes a link to
a third-party website that does a good job of explaining what the "average
age" of an exponentially smoothed series is. SqDist.md will need to be
modified at least once more when the OFR is complete, so that it can
link to the official USGS documentation in addition to all the supporting
documents and websites.