Added option to smooth SQ locally
By default, each element of the SQ vector (i.e., shat returned from additive())
is treated as if it is independent of its nearest neighbors in time. This is
not generally true, but to treat this in a completely statistically robust
manner would be computationally prohibitive. Instead, we apply a Gaussian
smoother/filter to a limited set of SQ elements in the neighborhood of the
current time step. The configuration parameter 'smooth' specifies the period
(in samples) at which one requires the amplitude to be attenuated by half.
Longer periods will be less attenuated, while shorter periods will be more
attenuated.

Also, added the running estimate of the standard error (i.e., sigmahat) to
the stream returned by process_one().