Pad to urlInterval in put_timeseries()
TimeseriesFactory.get_timeseries() pads to the requested interval with NaNs,
filling in start/end gaps. TimeseriesFactory.put_timeseries() is supposed to
do this (according to Jeremy Fee), but it was trimming NaNs from the start/
end before writing to file. This PR fixes this by using ObsPy's Stream.trim()
method, in a similar manner to TimesereisFactory.get_timeseries(), to pad to
the specified start and end of the urlInterval..