Add option to trim NaNs from SqDist input stream
The SqDistAlgorithm has had an issue since first being deployed wherein its
SV+SQ terms de-synchronized with the Dist term, which in turn led to frequent
reinitialization in real time operations. Reinitialization requires loading
and processing 90 days worth of data, which nullifies two major advantages of
this recursive algorithm (i.e., small data footprint, low cpu-usage).

This commit adds the ability to optionally trim NaNs off the end of the input
stream prior to calling the 'additive' Holt-Winters exponential smoother. This
ensures that the SV+SQ+Dist outputs are always synchronized, and that the
`next_starttime` element of the state file used to store the state between
calls to SqDistAlgorithm will always coincide with the beginning of the most
recent gap in a real time data stream.

Finally, this change does not address the situation where more than a single
"end gap" within a chunk of input data exists. In such a case, the algorithm
will still reinitialize because the `next_starttime` will not coincide with
the beginning of this gap. If this is undesirable, the calling procedure
should be set up such that the requested input chunks should be the same
length as the interval since the last call to SqDistAlgorithm. This can be
easily done given a recent commit to change the controller's `--realtime`
option.