Initial AdjustedValidate notebook
This notebook currently serves as an Algorithm Theoretical Basis,
validation report, and a not-so-user friendly tool for generating
Adjusted Data transform matrices manually. It needs to be cleaned,
and possibly broken into 2 or even 3 separate notebooks, but I'm
putting it into GitHub as a WIP to get some initial feedback.