Remove `import sys`
To satisfy flake8, and not trigger the F401 error, it was necessary to remove `import sys`, since it was no longer being used. This error slipped through Travis-CI because @erigler-usgs  got impatient waiting for Travis-CI to complete for a very simple PR and merged it before all tests were complete. Lesson learned.