Add SJT preliminary metadata to metadata.py
The actual calibration coefficints are filed in the GHSC, but due to
COVID-19 telework policies at the moment, it is not possible to go
and retrieve these. We've used nominal scaling values for the SJT
FGE fluxgate magnetometer here, providing only bare minimal info.
Once business returns to normal, this file should be updated for
SJT.