BRT calibrated values; add FDT to Metadata.py
Turns out we had calibration values for the BRW/BRT sensor in the
PCDCP system, so I used these instead of the default Narold values.
Also, @bgeels said FDT is ready to go online, so we added an entry
for that. FRD was using default Narod calibration values in PCDCP,
so we just kept those.