Pass unit tests again
The first attempt to fix TimeseriesUtility.pad_and_trim_trace() broke
a unit test. I'm not sure I agree with what the unit test defined as
passing, but I do understand the logic. This now satisfies the logic
of the unit test(s), while still fixing the floating point precision
problem to the default resolution of UTCDateTime objects (that is,
1e-6 or microseconds)